requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.16.1
flake8==7.1.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.2
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
samba==0.0.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uWSGI==2.0.26
virtualenv==20.26.5
whitenoise==6.7.0
